requirements.txt
# 基础依赖
setuptools
wheel

# Web框架
fastapi==0.103.1
uvicorn>=0.15.0
pydantic>=1.8.2
websockets>=10.0

# 环境配置
python-dotenv>=0.19.0

# 阿里云SDK
aliyun-python-sdk-core==2.15.1

# 数据处理
numpy>=1.20.0
matplotlib>=3.3.4

# 存储相关
oss2

# 类型检查
typing-extensions>=4.0.0

# 异步处理
aiohttp>=3.8.0
